create mock-notification-sys
